Flyttet ut configurasjonen i en egen fil
